moved reveal table values to hook.
